requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
pypdf
python-docx
langchain
pinecone
pdfplumber 
python-pptx
PyPDF2
python-multipart